DO NOT SUBMIT: limit lucid charge current during pd transition
When lucid is negotiating the transition to 20V, it is either unable to
send the ACk to negotiate after receiving PS_RDY from zinger or it iver
receives the PS_RDY from zinger. This is just a hack to limit the charge
current and prevent the negotiation issues.

BRANCH=none
BUG=chrome-os-partner:49659
TEST=test that lucid can charge normally

Change-Id: I732b8f205667ddfe733ac4765535aea7f5da35a9
Signed-off-by: Mary Ruthven <mruthven@chromium.org>