charger: Fix for boot from cut-off battery
Battery in cut-off mode wakes when voltage is applied to the PACK and
takes 2 to 3 seconds to initialize before capable of providing the power.
When the battery status changes from BP_CUT_OFF to BP_YES, charger input
current is set to board specific charger input current which is not
sufficient to boot the AP hence the system reboots. To avoid this issue,
added code to write charger manager negotiated current to charger input
current.

BRANCH=none
BUG=chrome-os-partner:49224
TEST=Manually tested on Kunimitsu.
     Used console command 'cutoff' to put the battery in cut-off mode.
     Inserted the adopter to wake the system, system doesn't reboot &
     the battery charges.

Change-Id: Ia5a1457506b4bef0b3dd27993e4b60ae64c8f746
Signed-off-by: Vijay Hiremath <vijay.p.hiremath@intel.com>