Update patch set 1
Patch Set 1: -Verified

Regarding shutting down on the G3 transition, there are some drawbacks. EC / PD will be back in RO, so boot may take longer, due to the need to verify / sysjump. I think it's reasonable to do PMIC shutdown after some G3 timeout.

Regarding the new callback vs existing, the existing callback is called as we're going to hibernate, and certain units of the chip have already been shutoff. The new callback is called before any actions are taken. We definitely cannot use the existing callback as-is for our purpose here since I2C won't work. We could try to move the existing callback earlier in the hibernate sequence, but it seems better to me to have distinct callbacks for different purposes.

Thinking about this more, I should add some error handling to the i2c write so we don't fail and end up in a while(1).

Patch-set: 1
Label: Verified=0