Update patch set 1
Patch Set 1:

Do you think it would be better to name it board_hibernate and then guarantee no return (execute instead of hibernate method, and maybe syshalt if it happens to illegally return)? It would be less flexible (if we ever want a board pre-hibernate that returns then we'll need to add something new or change it) but its purpose may be more clear.

Patch-set: 1