NPCX: Consolidate keyboard specific GPIO control
Previously the pin configuration (alternate function and GPIO setup) for
the keyboard scanning code was done in three places, hibernation state
was set in system.c, initial configuration was done in gpio.c and the
special case of inverting Column 2 was handled in keyboard_raw.c.  The
hibernation code in system.c was also ignoring the KEYBOARD_ROW/COLS
defines and forcing most, but not all, of the possible scan out pins
into a hibernation state.  That would be a problem for a board that
only uses some of the scan out pins for the keyboard matrix and others
as GPIOs.

Now all of the configuration and hibernation management is done in
keyboard_raw.c.  There is a new keyboard_raw function keyboard_hibernate
that is exposed just within the NPCX chip directory and is called from
the system hibernation code.  Hibernation in general could use a more
standard mechanism for drivers to manage the state change cleanly.  The
new hibernation code is careful to only switch pins to GPIO that are
actually being used by the keyboard matrix, as defined by the
KEYBOARD_ROW, KEYBOARD_COLS, and CONFIG_KEYBOARD_KSO_BASE defines.

Signed-off-by: Anton Staaf <robotboy@chromium.org>

BRANCH=None
BUG=None
TEST=make buildall -j

Change-Id: I48aba2ee8a0bed2eefd91c62f774e6ac357de394