Update patch set 2
Patch Set 2:

The problem is that the functions are also used elsewhere.
The inline declaration hints to the compiler (and it is just a hint) to inline power_get_pause_in_s5() within power/common.c.

But then we have power/baytrail.c (for example) using it, with only the header. The header states "this function should be inlined", the compiler responds "okay, maybe I would, but where is it?!?"

(and the same for the other functions)

Patch-set: 2