util: ec3po: Add OOBM queue and dynamic loglevels.
This commit adds an Out Of Band Managament queue which will allow the
console to receive commands outside of the PTY which it can take
action on.  The first use of this is to dynamically change the logging
level.  Prior to this change, changing the log level using dut-control
would not affect the log level of the console or interpreter.

BUG=None
BRANCH=None
TEST=Launch modified servod; issue dut-control loglevel:debug, verify
that debug messages from both servod and ec3po are emitted.  Then
issue dut-control loglevel:info and verify that no debug messages from
either are emitted.

Change-Id: I692824742b018da9540a81305985f6f355f716e6
Signed-off-by: Aseda Aboagye <aaboagye@google.com>