NPCX: Add chip specific GPIO APIs
These API's are specific to the NPCX and allow conversion to and from
the alternate function group/mask description of a pin to the GPIO
port/mask description.  The first consumer of these API's will be the
raw keyboard code that needs to be able to switch pins between being
GPIOs and alternate functions, so the keyboard scan out pins have
also been added to the gpio_alt_table.

Signed-off-by: Anton Staaf <robotboy@chromium.org>

BRANCH=None
BUG=None
TEST=make buildall -j

Change-Id: I0c1b112b3bbb195aefdc3e03039fcb8a3f80a581