Update patch set 1
Patch Set 1:

(1 comment)

> "gpio.h" is used only in chip level driver.
 > I suggest to rename it to gpio_chip.h.

Sure, my preference for includes is that they are done with a full relative path from the root of the project, but the EC codebase is not there, so we have to live with disambiguating include file names.

Given your other comment about the alt function table, I think this whole CL will go away and the portion of the table (and search function) will move to the keyboard_raw code.

Patch-set: 1