charger: Settings for charger voltage in no battery condition
In case of no battery condition, current code sets the charger voltage
to battery minimum voltage. For the reliability of the system and also
to avoid system reboot due to the voltage drop on VBATA, set the charger
voltage to the maximum voltage that the battery can support.

BUG=chrome-os-partner:50081
BRANCH=none
TEST=Manually tested on kunimitsu, removed the battery & then using EC
     console command 'charger', verified that the voltage is set to 8.7V.

Change-Id: I513b707e639301cda0ebd215f1a83cd78265975b
Signed-off-by: Vijay Hiremath <vijay.p.hiremath@intel.com>