mec1322: timer16: Ensure timer interrupt is disabled before timer
When the timer block is disabled and an interrupt fires simultaneously,
the block clock is gated off, so the interrupt may be impossible to
clear. Ensure that we disable our timer interrupt prior to disabling the
timer so that we don't fall into this interrupt storm case.

BUG=chrome-os-partner:48499
TEST=Pass 2000s of continuous calls to port_80_interrupt_disable() /
port_80_interrupt_enable() without WDT being triggered.
BRANCH=glados

Signed-off-by: Shawn Nematbakhsh <shawnn@chromium.org>
Change-Id: I6d073cb50122182bd3ad8cb7c20ed70089a0693b