Update patch set 2
Patch Set 2:

The port80_interrupt code is missing a

MEC1322_TMR32_STS = 1 ; */ Acknowledge Interrupt */

If my theory is correct, then we have an explanation of the storm and do not need the rest of this patch. We should on general principals also fix chip/mec1322/hwtimer.c:/__hw_clock_source_irq so it acknowledges the interrupt from the system clock timer as well as the event timer. The watchdog_check() already has this acknowledgement.

Patch-set: 2