cr50:signer: allow reading gnubby pin from a file
The need to enter the pin manually each time signing happens seems to
be an overkill.

With this patch the pin would be read from a file stored in a known
location, only if the file is readable by its user only. If the file
does not exits, has wrong permissions or contains a wrong pin - fall
back to manual pin entry.

BRANCH=none
BUG=chrome-os-partner:49950

TEST=verified that the process still works if the new file is not
     present, and manual pin entry is required in the following cases:
     - pin file is not present
     - pin file contents are wrong
     - pin file permissions are wrong

    verified that in case the file is present, it's access is limited
    to its owner and the contents are the correct pin, the signer
    proceeds to physical presence check without asking for the pin.

Change-Id: I84b4e24c9bc60f193540fc4ba06b4b0a88ddb6bf
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>