Update patch set 3
Patch Set 3: Code-Review-2

This we do not want.

The unlock protocol uses an ephemeral MAC key, which the application gets in exchange for the PIN.
That MAC key we cache in ~/.tmp/xxyyzz
That MAC key changes whenever the fob lost power.

We do not want to have a read-only leak of the filesystem yield the PIN to an observer.

Patch-set: 3
Label: Code-Review=-2