Update patch set 10
Patch Set 10:

Thank you very much for this update.
One question I have is about timestamps. Currently the EC keeps times with microsecond accuracy (except when it jumps backwards and forwards by about 1.3 hours :-( ). There is a special verb %T to print this out, and using it adds about 40 bytes to the stack usage. What is the plan for the future? Will EC-3PO be responsible for the timestamping? If so what accuracy are you expecting, and how will the latency - initially over the uart and later presumably over i2c or lpc (and maybe over ssh if you are running say ectool from the host) affect it? Maybe you plan to make every response packet from the host include a timestamp (perhaps just the lower 32 bits to keep the overhead down and expect ec-3po to handle the coarse timekeeping?).

Nice to see plans for adding colour to the output. Do you have any plans to be able to interleave the AP output in a different colour? Currently I do this with some emacs code and some people here find it very helpful.

Sorry to spam this bug, but I can't comment on the wiki or on the design doc.

Patch-set: 10