cr50: modify mpsse to work with the new device
The ultra debug board uses a different FTDI chip (vid:did 0403:6010),
which has two ports.

The SPI interface on this chip is hooked up to the second port, but
the code indiscriminately uses the first port when trying to open the
SPI channel, on all devices known to it.

Adding a new field in the supported_devices table allows to keep
default behavior, but use port B (the second port) when required.

Also, fix printout formatting problem.

BRANCH=none
BUG=chrome-os-partner:37754
TEST=tpmtest succeeds with the new board

Change-Id: I01f7937444c8df61d7439a66d9da89fb2cac5372
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>