Update patch set 6
Patch Set 6:

-- Also made the change in npcx - do you need it as a seperate patch ?

What you have now is fine, thanks.

-- I haven't used CONFIG_S0IX since the condition is enough to check in both cases. To remove the unwanted if condition in the S0IX not enabled case, something like the following can be done if needed. 

Yeah, I think it's best to change it as you suggested, even though the code is slightly messier. We want to skip that needless call on non-S0ix platforms.

-- This does open up a potential race due to asynchronous nature of hooks. If system transitions to S0, then the condition will not return true.

In fact, I think this should be perfectly safe, since hook_notify() is called synchronously from the chipset task (which itself updates the state variable). It would be unsafe if we used a deferred task or similar, though. LMK if you disagree.

Patch-set: 6
Label: Code-Review=0