isl9237: Increase Battery Disharging Current Prochot# threshold.
The default Battery discharging current Prochot# threshold value
for the Intersil 9237 charger is 4.096A which is low and prevents
CPU from going into turbo mode. Set it to the max value (12.800A)
so DCProchot# assertion is effectively disabled.

Note that DCProchot# assertion is only triggered when PSYS is
enabled which is the case in lars and kunimitsu.

BUG=none
BRANCH=glados
TEST=When lars is discharging from its two type-A and one type-C
USB ports, check that DCProchot# is not asserted.

Change-Id: I1643ada31456019e6a9c0df59985bf5dc053648c
Signed-off-by: Shamile Khan <shamile.khan@intel.com>