Update patch set 4
Patch Set 4:

Since Gerrit doesn't support giving just a subset of a CL a +2 it would be better to split something like this up into multiple CLs in the future.  In particular, having the USB-I2C bridge driver be a separate CL would be nice.  No need to do that this time, just an FYI for future CLs.

Patch-set: 4