Update patch set 1
Patch Set 1:

Also, your CL description -- "Add board-specific function to shutdown OS." Actually, it does much more than that -- it shuts off the power rails that power the EC. It results in considerable power savings. I do recommend that you implement this for lars, but for reasons that go beyond shutting down the OS (which seems to be a bug, as I just mentioned).

Patch-set: 1