jerry: Increase UART TX buffer size.
Found out that the string that is printed from console_init() doesn't
show up on the EC console for jerry.  It seems that perhaps due to the
priority of the console task and the numerous prints during boot, that
print never makes it to the TX buffer.  Currently, there's about 7200
bytes available in the shared memory region, which indicates that
there's still a lot of free space.  Therefore, increase the UART TX
buffer size.

BUG=None
BRANCH=None
TEST=Flash jerry and verify that 'console is enabled' string shows up on
the EC console.
TEST=make -j buildall tests

Change-Id: Id0603a3b758b5c600d0b59f27040ead2ce48bbaf
Signed-off-by: Aseda Aboagye <aaboagye@google.com>