Cr50: Debug: Add pinmux and gpiocfg commands
This adds two debugging commands to decode the PINMUX and GPIO
routings without having to look at a bunch of hex values.

They can easily be removed to save space, but they're kind of
handy for now.

BUG=chrome-os-partner:49952
BRANCH=none
TEST=make buildall, and test on Cr50

Run the "pinmux" and "gpiocfg" commands. Verify that the output
reflects the desired configs found in gpio.inc

Change-Id: I93b881bfd93dc100096bbd005a6c31b2669eda2f
Signed-off-by: Bill Richardson <wfrichar@chromium.org>