Cr50: Cleanup some of the GPIO handling
Just a little bit of code refactoring to make things a little
more like the way the other boards are configured.

BUG=none
BRANCH=none
TEST=make buildall, test on Cr50

No functional changes, so nothing new to test. However, with this
CL, declaring a Cr50 interrupt that triggers on both rising and
falling edges will fail to compile, since that isn't supported on
the SoC.

Change-Id: Icb0c8203544c580e726ad2ca3dcd799042d3a451
Signed-off-by: Bill Richardson <wfrichar@chromium.org>