Update patch set 1
Patch Set 1: Code-Review+1

(2 comments)

As I'm sure you are aware, this CL is almost identical to https://chromium-review.googlesource.com/#/c/327407/. Does it make sense to move this code (and some other related common code) to some common library so it doesn't have to be implemented 3 times in flashrom / mosys / ectool? Maybe it's something to think about for the future.

Patch-set: 1
Label: Code-Review=+1