Update patch set 2
Patch Set 2:

(2 comments)

Based the description of the problem ("PI3USB9281 may not assert the interrupt line if device is detached and reading the interrupt register simultaneously"), I don't think this change will completely eliminate the bug, but I'm ok with landing it. Hopefully we can implement the GPIO changes I suggested on the crbug and then stop relying on the contents of the interrupt register altogether (always do redetection on interrupt).

Patch-set: 2
Label: Code-Review=0