Update patch set 2
Patch Set 2:

Many years ago we had thought we might make a library (or libraries?) to be shared among a few tools. The flashmap library is as far as that idea went AFAIK. Since this functionality should (hopefully) not change again it's probably not worth doing just for the locks.

Patch-set: 2