Update patch set 1
Patch Set 1: Verified+1 Trybot-Ready+1

Hey all, I've added a larger than seems warranted number of reviewers to this CL because I'm doing something that I would like us to do more and want feedback on it, that is explicitly using a full path from the base of the project when including a header file.  Dumping all the headers into include and relying on compiler search paths to find the header you mean is leading to people putting inappropriate contents in the supposedly common/generic headers.  This function is a good example.

Patch-set: 1
Label: Trybot-Ready=+1
Label: Verified=+1