Update patch set 1
Patch Set 1:

We seem to have a semi-standard way of dealing with duplicate header files names in the chip folder -- add "_chip" to the name, eg. config_chip.h, pwm_chip.h, adc_chip.h, etc. Should we continue following that, or shall we change our headers + includes in chip/ to use the full path?

I'm slightly biased toward keeping the scheme as-is to reduce churn, but I'm ok with either.

Patch-set: 1