Update patch set 1
Patch Set 1:

> We seem to have a semi-standard way of dealing with duplicate
 > header files names in the chip folder -- add "_chip" to the name,
 > eg. config_chip.h, pwm_chip.h, adc_chip.h, etc. Should we continue
 > following that, or shall we change our headers + includes in chip/
 > to use the full path?
 > 
 > I'm slightly biased toward keeping the scheme as-is to reduce
 > churn, but I'm ok with either.

Yeah, my problem with that solution is that it doesn't scale very well.  It makes it hard for chip dependent code to include exactly the file it needs, and if there are family variant versions of some headers then you're adding the family name into the header name as well.  I agree that it would be some churn, but it doesn't have to happen all at once, just as files are touched or moved around.  Personally, if I could I would retroactively remove the default behavior of searching the current directory for headers as well.  I've found default search paths (well defaults in almost any situation) to be problematic, and prefer to be explicit instead.

Patch-set: 1