mec1322: Do not shutdown LPC in deepsleep.
During the resume sequence of S0ix EC can receive host commands early
in the resume path when LPC is still disabled in EC. Host messages
will be lost if the LPC interface with the kernel is down.

Clock control was programed to 2 which means ring oscillator is
shut down after completion of everty LPC transaction.To restart
the oscillator EC should enable a wake interrupt on LPC LFRAME number
and this mode can cause an increase in the time to
respond to the LPC transactions.

Keeping LPC always on shows minimal power impact as per datasheet
Pg.390. The impact is < 0.45mW.

BUG=chrome-os-partner:50627
TEST=Enter into S0ix and exit reliably.
BUILD=None

Change-Id: I670b9b45c3a85c9bca249312a73a25dca52b313a
Signed-off-by: Divya Jyothi <divya.jyothi@intel.com>