mec1322: Do not shutdown LPC in deepsleep
During S0ix EC can receive Host commands anytime and if LPC is
shut down by EC we will not be able to receive them.

Clock control was programed to 2 which means ring oscillator is
shut down after completion of everty LPC transaction.To restart
the oscillator EC should enable a wake interrupt on LPC LFRAME number
and this mode can cause an increase in the time to
respond to the LPC transactions.

Keeping LPC always on shows minimal power impact as per datasheet
Pg.390. The impact is < 0.45mW.

BUG=chrome-os-partner:50627
TEST=Enter into S0ix and exit reliably.
BUILD=None

Change-Id: I670b9b45c3a85c9bca249312a73a25dca52b313a
Signed-off-by: Divya Jyothi <divya.jyothi@intel.com>