kunimitsu: Disable wake up from Keyboard or trackpad in tablet mode
When system is in S3 and in tablet mode, disable key scanning and
power down trackpad so that the system does not resume from sleep
when these periperals are used.

BUG=none
BRANCH=glados
TEST=Manual on kunimitsu. Boot system with lid open in clamshell
mode. Open a shell window and type powerd_dbus_suspend to bring
the system to suspend state. Now fold the system to tablet mode.
Verify that pressing a key or touching the trackpad does not
cause system to wake up.

Change-Id: I0e58c154fa7a32211f6811fe598dd43e1da6e895
Signed-off-by: Shamile Khan <shamile.khan@intel.com>