Update patch set 1
Patch Set 1:

> (1 comment)
 > 
 > can you please explain the need for key reverting, why does the bn
 > library representation not match the tpm2 library it was written to
 > work with?

It's convention for keys to be serialized in big-endian format.  e.g. this is how openssl exports / imports keys (BN_bn2bin and BN_bin2bn).  

If the underlying hardware is of differing endianness, as is the case here, a conversion step is necessary.  That's what reverse does.  I could call it bin2bn if you prefer.

Patch-set: 1