TEST-ONLY: elm: Add elm changes
This adds changes compared to Oak rev5, including:
  - motion sensors are two KX022, BMI160 is removed
  - ALS sensor is removed
  - LEDs are two bi-color LED (blue and orange)
  - Add EN_TP_INT_L to enable/disable touchpad interrupt
  - pi3usb30532 is removed (need driver for 7688?)

This also removes codes using system_get_board_version(), and a little
bit of coding style.

Note that ANX7688 related changes are not here.

Battery setting is kept to support 2S, and this should be modified
when we don't need to support 2S anymore.

BRANCH=none
BUG=none
TEST=make BOARD=elm -j

Change-Id: Ibe6b47824814684b9bfb53bf52b152e0e1b5550b
Signed-off-by: Koro Chen <koro.chen@mediatek.com>