Update patch set 1
Patch Set 1:

Yes, it is doable assuming both G781 and G782 will not be used together on the same board.

In G781.h, I will create a common name G78X_xxxxxxx that exist on both chip and use #ifdef to select which.

In G781.c, I will use #ifdef for the extra sensor and the extra status register.

Will this work better?

Patch-set: 1
Label: Code-Review=0