Skylake: service host events triggered after  ACPI_QUERY_EVENT
On S3 entry all pending events are cleared from coreboot
SMI handler. If a host event is triggered after this, but
before SLP_S3 gets asserted, the event doesn't get cleared.
Clear the pending events from S0S3. Also if the lid is open
and there is a pending lid event, clear it and re-trigger
the event to resume from S3

BUG=None
BRANCH=glados
TEST=System should wakeup by lidopen after back to back lidopen
and close.

Change-Id: Ia14c02a3f05710a8deb5b3510c6fc074bfa82c03
Signed-off-by: Jenny TC <jenny.tc@intel.com>
Signed-off-by: Divagar Mohandass <divagar.mohandass@intel.com>