common: add forced lid open function
An earlier patch added a host command for forced lid
open for a factory use case. This rafactors the code
to add a common function which is available to other
functions.

In certain scenarios where we need to wake
up the system in a lid closed scenario due to an error
condition, we need to do a forced lid open.

BRANCH=Glados
BUG=chrome-os-partner:51441
TEST=Build for Lars

Signed-off-by: Archana Patni <archana.patni@intel.com>
Signed-off-by: Subramony Sesha <subramony.sesha@intel.com>

Change-Id: Ibc1a44b4704f4f8942cab815a896be273c93c92a