Update patch set 3
Patch Set 3:

I was also interested in the host command throttling options. I am not sure how the current values are chosen. In particular, what does a CONFIG_HOSTCMD_RATE_LIMITING_RECESS value of 30 ms translate to? Is it enough for the scheduler to run one lower priority task before returning to HOSTCMD task?
It may be that the current settings of 500 and 30 are OK for a busy flash operation when there is only one low priority task in between HOSTCMD and HOOK tasks but problematic now that we have two extra tasks for type-C.

Patch-set: 3