Update patch set 2
Patch Set 2:

There is a pwm_fan_resume() in common/fan.c that looks like it should do this with the same resume hook.  Is that hook not working?

It looks like sentry is half-enabled for DPTF fan control, what is the plan there?  No matter what the decision is it should be cleaned up in coreboot so it is clear what is actually controlling the fan.

Please file an issue and set the BUG= line so we can ensure the discussion and decision is tracked in the bug database.

Patch-set: 2
Reviewer: Gerrit User 1000100 <1000100@3ce6091f-6c88-37e8-8c75-72f92ae8dfba>
Label: Trybot-Ready=0