Update patch set 1
Patch Set 1:

Todd in your comment in crosbug:50627 you have mentioned 
"Userland fix is simply to not enable suspend-to-idle in powerd.  That has already been implemented (never enabled in M50 for lars/kuni) and should handle all but the s0i0 SLP_S0 assertions."
Not enabling suspend-to-idle will not handle S0i0 SLP_S0 assertions. Is that correct? Then Fan and LED will be a bad user experience for lars. Please see the video attached in crosbug: 50790.

Patch-set: 1