common: introduce new clear_and_set_host_event function and switch lidopen/close
In certain conditions, a host event may not have the desired end result and the
host event might be repeated by the user resulting in an accumulation of events.
For e.g. if a previous lid open is not cleared and a lidclose/lid open sequence
is repeated, this can lead to an accumulation of events.

This patch introduces a new clear/set function which will clear any preceding
event of the same type and then set a new version.

BRANCH=Glados
BUG=chrome-os-partner:51613
TEST=Do repeated test of lidclose/lidopen and check if there is any accumulation
of events through hostevents interface.

Change-Id: I6a2054362ca32d7d1683c37e3d4b101b4c1795b7
Signed-off-by: Archana Patni <archana.patni@intel.com>
Signed-off-by: Subramony Sesha <subramony.sesha@intel.com>